Html수정

+++ security1/src/main/java/com/cos/security1/repository/PagingRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PagingRepository extends JpaRepository<Board, Integer> {
+//게시글 페이징처리 및 검색 기능
+public interface PagingRepository extends JpaRepository<Board, Long> {
     Page<Board> findByTitleContaining(String searchKeyword, Pageable pageable);
 }
